Adding dropbear SSH server
Its main purpose is to serve as an interactive entrypoint for the
container. Moreover, a special patched version of dropbear is used,
which is tuned to switch login shell automatically from fish back
to bash, if _vscode suffix is added into the username. This is
needed to workaround the poor quality of VSCode Remote SSH plugin,
which works with bash shell only.